requirements.txt
torch
transformers
Pillow
click
tqdm
numpy
scipy
